% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_clean_objects.R
\name{collect_clean_objects}
\alias{collect_clean_objects}
\title{Collect the objects of a cleaning workflow from a targets store}
\usage{
collect_clean_objects(
  store,
  prefix = "bio_",
  excludes = c("path", "dir", "file", "env", "rich", "names"),
  extras = c("clean_end")
)
}
\arguments{
\item{store}{Path to a targets store}

\item{prefix}{Object name prefixes. Objects within \code{store}
that match this will be collected}

\item{excludes}{Any object that matches any regex provided in \code{excludes} will
not be collected}

\item{extras}{Any object that matches any regex provided in \code{extras} will be
collected}
}
\value{
tibble
}
\description{
Collect the objects of a cleaning workflow from a targets store
}
