% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_toi.R
\name{collect_toi}
\alias{collect_toi}
\title{Create a vector of taxa of interest (toi) from yaml files}
\usage{
collect_toi(
  dir = "settings",
  toi_regex = "\\\\/toi",
  exclude = NULL,
  include = NULL,
  keep_previous = TRUE,
  store_path
)
}
\arguments{
\item{dir}{Directory in which \code{toi_regex} files will be found}

\item{toi_regex}{Regular expression that will pick up any yaml files
containing the taxa of interest.}

\item{exclude}{Character. Vector of taxa used to limit the resulting
vector. i.e. only taxa that do not appear in \code{exlude} will make it into
the resulting vector.}

\item{include}{Character. Vector of taxa used to limit the resulting
vector. i.e. only taxa that appear in \code{include} will make it into the
resulting vector.}
}
\value{
Character vector
}
\description{
Usually used to create a vector over which tarchetypes::tar_map is run
}
