% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_values.R
\name{collect_values}
\alias{collect_values}
\title{Create a vector of values from a targets store object}
\usage{
collect_values(
  dir,
  object,
  column,
  collect_col,
  exclude = NULL,
  include = NULL
)
}
\arguments{
\item{dir}{Directory containing targets store(s).}

\item{object}{Name of object within store(s).}

\item{collect_col}{Character name of column within objects found in store(s)
with the values to collect.}

\item{exclude}{Character. Vector of values used to limit the resulting
vector. i.e. only values that do not appear in \code{exclude} will make it into
the resulting vector.}

\item{include}{Character. Vector of values used to limit the resulting
vector. i.e. only taxa that appear in \code{include} will make it into the
resulting vector.}
}
\value{
Character vector
}
\description{
Usually used to create a vector for dynamic or static branching
}
