% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tiles.R
\name{combine_tiles}
\alias{combine_tiles}
\title{Combine tiles into a single raster}
\usage{
combine_tiles(tiles, out_file, ...)
}
\arguments{
\item{tiles}{Paths to tiles, possibly from
\code{predict_terra_tiles()}, or list of \code{SpatRaster} tiles.}

\item{out_file}{Character. Name of the saved file.}

\item{...}{Passed to \code{terra::merge()}. e.g. \code{overwrite = TRUE}; or
\code{wopt = list(datatype = "INT1U")}.}
}
\value{
\code{out_file}
}
\description{
Combine tiles into a single raster
}
