% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_tar_group.R
\name{env_tar_group}
\alias{env_tar_group}
\title{Prepare a target for batched branching}
\usage{
env_tar_group(
  df,
  column = "n",
  batches = envFunc::use_cores(absolute_max =
    yaml::read_yaml("settings/setup.yaml")$max_cores)
)
}
\arguments{
\item{df}{dataframe}

\item{column}{column to use when assigning rows to groups}

\item{batches}{How many groups to create?}
}
\value{
\code{targets::tar_group()}ed dataframe
}
\description{
Helps to minimise long tails while processing batched data frames by taking
into account one of the columns (\code{column}) within the dataframe (usually \code{n})
that is likely to be correlated with processing time. Remember to set
\code{iteration = "group"} within the target.
}
