% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_min_date.R
\name{extract_min_date}
\alias{extract_min_date}
\title{Extract a minimum date to use in the project, based on the settings}
\usage{
extract_min_date(
  settings,
  temporal_extent = "ext_time",
  max_year = as.numeric(format(Sys.Date() - 365, "\%Y"))
)
}
\arguments{
\item{settings}{list. usually from yaml::read_yaml("settings/setup.yaml")}

\item{temporal_extent}{Character. Name of the temporal extent within
\code{settings}}

\item{max_year}{Numeric. Year. e.g. 2024. Default is the last full year via
\code{as.numeric(format(Sys.Date() - 365, "\%Y"))}}
}
\value{
Character in date format. e.g. 1970-01-01
}
\description{
Extract a minimum date to use in the project, based on the settings
}
