% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_temporal_grain.R
\name{extract_temporal_grain}
\alias{extract_temporal_grain}
\title{Get the temporal grain from settings}
\usage{
extract_temporal_grain(settings)
}
\arguments{
\item{settings}{list. usually from yaml::read_yaml("settings/setup.yaml")}
}
\value{
Character. "year", "month" or "date"
}
\description{
Get the temporal grain from settings
}
