% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_context_combos.R
\name{find_context_combos}
\alias{find_context_combos}
\title{Find context combinations based on settings}
\usage{
find_context_combos(
  proj,
  store,
  settings = yaml::read_yaml("settings/scales.yaml")$default,
  ext = settings$extent$ext_time,
  lev = settings$aoi$filt_lev,
  tax_grains = settings$grain$taxonomic,
  lev_all_type = "all_in_vec",
  stop_if_not_run = TRUE,
  aoi_setting = "aoi",
  track_file
)
}
\arguments{
\item{proj, store}{Project and store names used to specify the directory path to look for existing stores & files
relating to the context combos.}

\item{settings}{List of contexts (or 'scales') usually from yaml::read_yaml("settings/scales.yaml").
Must contain extent, grain & optionally aoi as first list elements, with secondary lists of
vector, filt_col, filt_level, buffer, ext_time, region_taxa under extent,
res_x, res_y, res_time, taxonomic under grain, and
vector, filt_col, filt_level, buffer under aoi.}

\item{ext}{Character vector of temporal extents to vary the \code{ext_time} setting in \code{settings$extent},
e.g. c("P10Y", "P20Y", "P30Y", "P50Y", "P100Y"). Numbers must be preceded with 'P' and followed by 'Y'.}

\item{lev}{Character vector of filter levels to vary the \code{filt_level} setting in \code{settings$aoi}.
As with \code{ext} & \code{lev}, can be specific levels or 'all' to run all available (see \code{lev_all_type} below for 'all' options).}

\item{tax_grains}{Character vector of filter levels to vary the \code{taxonomic} setting in \code{settings$grain},
e.g. c("species", "subspecies").}

\item{lev_all_type}{Type of method used to determine 'all' values for \code{lev}.
Either, 'all_in_vec' which retrieves all levels in the column in the vector corresponding to those specified by
\code{settings$aoi$filt_col} and \code{settings$aoi$vector} respectively,
or 'already_run', which retrieves levels from existing store directories that have already been run,
i.e. those that are available.}

\item{stop_if_not_run}{Logical. Stop if the any of the specified context cominations haven't been run
and don't have a \code{track_file} (see below)?}

\item{aoi_setting}{Character. Name of the 'aoi' type setting containing the list of aoi related measures
(vector, filt_col, filt_level, buffer), e.g. sometimes referred to as 'region' instead of 'aoi'.}

\item{track_file}{Character name of file to use for tracking if a store relating to a context combo has been run.
Usually one of the last files created in the project/store, and or one used downstream.}
}
\value{
Data frame with ext_time, rank, filt_level, path, exists and combo columns,
with each row representing a combination of the ext_time, taxonomic & filt_level inputs.
}
\description{
Finds all context combinations for mulitple ext_time, taxonmic grain, and filt_level values
specified as vectors in a yaml file. Also returns the associated store paths for each combo via envFunc::name_env_out
corresponding to a file in the store, and an indication if it exists.
}
