% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_context_files.R
\name{find_context_files}
\alias{find_context_files}
\title{Find context files based on settings combos in a data frame}
\usage{
find_context_files(
  project = basename(here::here()),
  scales_yaml = "scales.yaml",
  combos_df,
  track_file = "_targets.yaml",
  stop_if_not_run = TRUE,
  store_base = "../../out"
)
}
\arguments{
\item{project}{Character. One or more project names. e.g. basename(here::here()) from
inside, say, an RStudio project.}

\item{scales_yaml}{Character. Name of the \code{scales.yaml} file to obtain context values for
non-varied context settings.}

\item{combos_df}{Data frame of settings context combinations for all the settings contexts, where the columns
are the settings and the rows are each combination of different values for each setting, e.g. as produced by
envTargets::make_context_combos.}

\item{track_file}{Name of file with extension to use for tracking if a store relating to a context combo has been
run. Usually one of the last files created in the project/store, and or one used downstream.
If in a store/folder below the lowest context specified in \code{combos_df}, then include
the store path with the file, e.g. "reg_cont/objects/reg_cont_tbl_tidy".}

\item{stop_if_not_run}{Logical. Stop if the any of the specified context files haven't been run
and don't have a \code{track_file}?}

\item{store_base}{Path to base directory of project store, e.g. "../../out"  or "/projects/data".}
}
\value{
Data frame with extent_time, rank, filt_level, path, exists and combo columns,
with each row representing a combination of the extent_time, taxonomic & filt_level inputs.
}
\description{
Find context files based on settings combos in a data frame
}
