% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file.R
\name{find_file}
\alias{find_file}
\title{Search down through a directory until file is found}
\usage{
find_file(path = "../../out", find = "_targets.yaml", recurse_depth = 4)
}
\arguments{
\item{path}{path to search}

\item{find}{file to find}

\item{recurse_depth}{maximum value passed to the recurse argument of
\code{fs::dir_ls()}}
}
\value{
Character
}
\description{
Search down through a directory until file is found
}
\keyword{internal}
