% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_log_errors.R
\name{find_log_errors}
\alias{find_log_errors}
\title{Find any errors (or warnings) from a targets workflow}
\usage{
find_log_errors(store, days = NULL, warn = FALSE, err = TRUE)
}
\arguments{
\item{store}{Store in which the logs are being saved}

\item{days}{Numeric. To limit results to only the last \code{days} days.}

\item{warn}{Logical. Search for warnings?}

\item{err}{Logical. Search for errors?}
}
\description{
i.e. from the logs generated by setting
\code{options_local = crew_options_local(log_directory = blah)} in
\code{crew::crew_controller_local()}
}
