% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deps.R
\name{get_deps}
\alias{get_deps}
\title{Get dependencies

For use in \code{render_with_deps()}}
\usage{
get_deps()
}
\description{
Get dependencies

For use in \code{render_with_deps()}
}
\author{
Shir Dekel (modified by nw)
}
