% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tar_meta.R
\name{get_tar_meta}
\alias{get_tar_meta}
\title{View targets metadata with progress}
\usage{
get_tar_meta(store, incl_branches = FALSE, keep_cols = NULL)
}
\arguments{
\item{store}{Path to the store to check (usually tars$blah$store)}

\item{incl_branches}{Include targets with "_branches" in the name (ie dynamic branches)? Default FALSE can greatly reduce table length.}

\item{keep_cols}{Extra columns to include from tar_meta.}
}
\value{
A data frame with names and metadata of targets. Always returns name, date, time, seconds (runtime), size (with useful units), format, and warnings; add more with \code{keep_cols}.
}
\description{
\code{targets::tar_progress} only returns target names and status (skipped, completed, etc); \code{targets::tar_meta} returns
more metadata but no progress. This function combines the two and formats some columns to be more useful
(e.g. more human-readable sizes, with sensible units; separate date and time)
}
\note{
Previously in envShiny.
}
\seealso{
\code{\link[envShiny:tar_meta_opts]{envShiny::tar_meta_opts}} in the options argument of \code{DT::datatable()} is useful for easy consistent output formatting in Shiny metadata tabs.
}
