% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_context_combos.R
\name{make_context_combos}
\alias{make_context_combos}
\title{Make context combinations data frame based on settings}
\usage{
make_context_combos(
  project = basename(here::here()),
  scales_yaml = "scales.yaml",
  vars,
  all_vec_lev = FALSE,
  vec_dir = fs::path(yaml::read_yaml("settings/setup.yaml")$data_dir, "vector")
)
}
\arguments{
\item{project}{Character. One or more project names. e.g. basename(here::here()) from
inside, say, an RStudio project.}

\item{scales_yaml}{Character. Name of the \code{scales.yaml} file containing the \code{project} settings.}

\item{vars}{List of settings and values to vary, e.g. list(extent_time = c("P10Y", "P50Y", "P100Y")
, taxonomic = c("species", "subspecies")).}

\item{all_vec_lev}{Logical. Use all levels in the spatial vector specified by the aoi_vector setting within the
column specified by the aoi_filt_col setting in the \code{scales_yaml}? Used instead of specifying a vector in \code{vars}
for the aoi_filt_level setting. Useful where there are large numbers of aoi_filt_levels.}

\item{vec_dir}{Vector directory containing the vector file that corresponds to the aoi_vector setting for the
\code{project} in the \code{scales_yaml}. Only used if \code{all_vec_lev} is TRUE, for retreiving all the levels in the filt_col
in the specified vector. Currently has to be a geoparquet with ".parquet" extension accessible by
sfarrow::st_read_parquet.}
}
\value{
Data frame with all possible combinations of the variables provided, along with
fields with constant values for the other non-varied settings contexts.
}
\description{
Make context combinations data frame based on settings
}
