% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tar_id.R
\name{make_tar_id}
\alias{make_tar_id}
\title{Make a suitable targets 'id'}
\usage{
make_tar_id(char, replacement = "_")
}
\arguments{
\item{char}{Character}

\item{replacement}{Character. What to use to replace punctuation and spaces?}
}
\value{
Character with punctuation or blanks replaced with \code{replacement}
}
\description{
Make a suitable targets 'id'
}
