% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tile_extents.R
\name{make_tile_extents}
\alias{make_tile_extents}
\title{Make tile (extents) for raster split, apply, combine}
\usage{
make_tile_extents(
  base_grid_path,
  aoi = NULL,
  tile_name = "tile__",
  tile_length = 512,
  tile_size = 1e+05,
  min_tiles = 4,
  add_hectares = TRUE
)
}
\arguments{
\item{base_grid_path}{Character. Path to a template raster.}

\item{aoi}{Optional sf object. Used to define the outer extent of tiles'}

\item{tile_name}{Character. Prefix for the (file) name of each tile.}

\item{tile_length}{Numeric. Number of rows and columns in each tile.}

\item{tile_size}{Numeric. Ignored if tile_length is not \code{NULL}. Tile size in
units of the raster at \code{base_grid_path}. e.g. with \code{tile_size} of \code{10000}
a projected raster with units of metres will result in tiles of size 10 km by
10 km.}

\item{min_tiles}{Numeric. Ignored if tile_length is not \code{NULL}. Minimum
number of tiles to return. If \code{tile_size} returns less than \code{min_tiles}, then
\code{min_tiles} will be used instead of \code{tile_size}.}

\item{add_hectares}{Logical. Append the hectares of each tile? Ignored if
the raster at \code{base_grid_path} is not projected.}
}
\value{
tibble of extents
}
\description{
Make tile (extents) for raster split, apply, combine
}
