% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_store_metadata.R
\name{parse_store_metadata}
\alias{parse_store_metadata}
\title{Parse metadata in the file path of a (env) targets store}
\usage{
parse_store_metadata(
  project = basename(here::here()),
  store_base = "../../out",
  targets_yaml = "_targets.yaml",
  scales_yaml = "scales.yaml"
)
}
\arguments{
\item{project}{Character. Project name. e.g. basename(here::here()) from
inside, say, an RStudio project.}

\item{store_base}{Character. Path to the level above the project store.}

\item{targets_yaml}{Character. Name of the \verb{_targets.yaml} file(s) to search
for within the \code{store_base}}

\item{scales_yaml}{Character. Name of the \code{scales.yaml} file(s) to search
for within the \code{store_base}}
}
\value{
Tibble of parsed metadata
}
\description{
Parse metadata in the file path of a (env) targets store
}
