% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_terra_tiles.R
\name{predict_terra_tiles}
\alias{predict_terra_tiles}
\title{Predict a model across tiles}
\usage{
predict_terra_tiles(
  extent_df,
  predictors,
  model,
  model_read_fun = readRDS,
  terra_options = NULL,
  out_dir,
  force_new = TRUE,
  load_packages = NULL,
  check_tifs = TRUE,
  ...
)
}
\arguments{
\item{extent_df}{Tibble from \code{make_tile_extents()}.}

\item{predictors}{Character. Paths to predictor rasters.}

\item{model}{Either model object, or path to model, with a predict method.}

\item{model_read_fun}{If \code{model} is a path, what function to use to read the
model. Defaults to \code{readRDS()}.}

\item{terra_options}{List of options to be passed to \code{terra::terraOptions()}.
e.g. \code{list(memfrac = 0.5)}.}

\item{out_dir}{Character. Name of directory into which predicted tiles will
be saved.}

\item{force_new}{Logical. If TRUE, predict tile even if it already exists.
Leave as \code{TRUE} to let targets decide if the tile needs to be rewritten.}

\item{load_packages}{Character. Any packages to load before predict.}

\item{check_tifs}{Logical. If \code{TRUE} and \code{force_new} is \code{FALSE}, an attempt
will be made to open any existing tile. If that fails, the tail will be
deleted. While expensive in time, this ensures that any crashes that result
in incomplete tiles are fixed.}

\item{...}{Passed to \code{terra::predict()}. Object, model and filename arguments
are already passed so are not needed here. e.g. \code{na.rm = TRUE};
\code{overwrite = TRUE}; or \code{wopt = list(datatype = "INT1U")}}
}
\value{
Path to the predicted tile(s). Writes tile into \code{out_dir}.
}
\description{
Predict a model across tiles
}
