% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bookdown_yaml.R
\name{prepare_bookdown_yaml}
\alias{prepare_bookdown_yaml}
\title{Prepare _bookdown.yaml within a targets workflow}
\usage{
prepare_bookdown_yaml(
  report_dir = "report",
  store = tars$report$store,
  output_dir = "compiled_report",
  repo = usethis::git_remotes()$origin
)
}
\arguments{
\item{report_dir}{Character. Where should the \verb{_bookdown.yaml} file be saved?}

\item{store}{Character. Path to the targets store into which the report
should be saved}

\item{output_dir}{Character. Name to give the folder in \code{store} where the
report will be saved.}
}
\value{
Path to saved \verb{_bookdown.yaml}.
}
\description{
Prepare _bookdown.yaml within a targets workflow
}
