% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_env.R
\name{prepare_env}
\alias{prepare_env}
\title{Prepare a tibble of environmental layers from settings}
\usage{
prepare_env(set_list, reg_exp = "\\\\.tif", ...)
}
\arguments{
\item{set_list}{list. usually from yaml::read_yaml("settings/setup.yaml")}

\item{reg_exp}{Character. Used to limit returned files.}

\item{...}{Passed to envRaster::name_env_tif}
}
\value{
tibble of raster paths and meta data parsed into columns
}
\description{
'Seasons' functionality removed 2025-09-05. If seasons are needed, left join
the result of \code{prepare_env()} to the result of \verb{make_seasons()``$seasons}.
}
