% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_output_yaml.R
\name{prepare_output_yaml}
\alias{prepare_output_yaml}
\title{Write _output.yaml for html and docx reports}
\usage{
prepare_output_yaml()
}
\value{
File paths (relative) to _output.yaml and style file (for bs4_book output).
}
\description{
To select output format, use \code{output_format = "bookdown::bs4_book"} / \code{"bookdown::word_document2"} in \code{\link[=render_with_deps]{render_with_deps()}}.
}
