% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_with_deps.R
\name{render_with_deps}
\alias{render_with_deps}
\title{Render bookdown and force Rmd file dependencies}
\usage{
render_with_deps(
  input_directory = ".",
  output_directory = NULL,
  deps = NULL,
  ...
)
}
\arguments{
\item{input_directory}{Path in which \code{index.Rmd} can be found from \code{here::here()}.
Will almost certainly fail if \code{input_directory} is not found within \code{here::here()}.}

\item{output_directory}{Path (often a tars store) in which final output(s)
will be found. If \code{NULL} (default) \code{output_directory} will be
\code{yaml::read_yaml(here::here(input_directory, "_bookdown.yaml"))$output_directory} if that
\code{bookdown.yaml} exists, otherwise \code{here::here(input_directory, "_book")}.}

\item{deps}{Any dependencies for the report. Note that these are not made
available within the knit, they just trigger the target to run.}

\item{...}{Arguments to boookdown::render_book, especially for output_format.}
}
\value{
Return value is just the \code{output_directory}. Files needed
for the book are written to the output directory.
}
\description{
Moving the report to the targets store relies on output_directory being
specified, usually via the _bookdown.yaml file, itself usually written within
report_prep.R using \code{envTargets::prepare_bookdown_yaml()}.
}
\details{
Enables a bookdown 'target' within a \href{https://books.ropensci.org/targets/}{targets workflow}.
I was unable to get \href{https://mps9506.r-universe.dev/bookdowntargets/doc/manual.html}{bookdowntargets}
to work without error. That may be a better way to achieve this. Currently
(20250520) \code{deps} still need to be provided inside bookdown scripts (i.e. providing
\code{deps} to the function simply tells targets when the book needs to be remade,
it does not make those \code{deps} available inside the knit.)

Based on code found here: https://github.com/shirdekel/phd_thesis/commit/cc66e1b9d9305a4b21e8489836545ecc4475f9ee
}
\author{
Shir Dekel (modified by nw)
}
