% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_run_all.R, R/run_all.R
\name{run_all}
\alias{run_all}
\title{Run all targets stores for multiple settings contexts}
\usage{
run_all(
  scales_yaml = "scales.yaml",
  combos_df,
  current_proj = basename(here::here()),
  upstream_proj,
  upstream_track_file = "_targets.yaml",
  current_track_file = "_targets.yaml",
  force_new = TRUE,
  run_file = "run.R",
  current_store_base = fs::path("..", "..", "out"),
  upstream_store_base = fs::path("..", "..", "out")
)

run_all(
  scales_yaml = "scales.yaml",
  combos_df,
  current_proj = basename(here::here()),
  upstream_proj,
  upstream_track_file = "_targets.yaml",
  current_track_file = "_targets.yaml",
  force_new = TRUE,
  run_file = "run.R",
  current_store_base = fs::path("..", "..", "out"),
  upstream_store_base = fs::path("..", "..", "out")
)
}
\arguments{
\item{scales_yaml}{Character. Name of the \code{scales.yaml} file to obtain settings contexts.
Must be the same in both the current and upstream projects.}

\item{combos_df}{Data frame of combinations of the settings to be varied, where the columns are
the setting names and the rows are each combination of different values for each setting,
e.g. as produced by envTargets::make_context_combos. The setting names must correspond to those in
the settings for the \code{current_proj} element in the \code{scales_yaml}.}

\item{current_proj, current_store}{Current project & store names in which running this function and
running multiple outputs, e.g. 'envRegCont' & 'reg_cont'. Used for checking if existing \code{track_file}s exist.}

\item{upstream_proj, upstream_store}{Upstream project & store names required for the current project,
i.e. the precursor project that contains the combos of settings context directories & the store within them,
e.g. 'envRange' & 'grd'.}

\item{current_track_file, upstream_track_file}{Names of files to use for tracking if a store
relating to a context combo has been run. Usually one of the last files created in the project/store,
and or one used downstream, e.g. 'grd_files' in the envRange project as the upstream file, and
'reg_cont_tbl_tidy' in envRegCont if that project was the current. Otherwise, defaults to the
'_targets.yaml', as an indication the store has been run (although possibly not fully).}

\item{force_new}{Logical. Force new runs of stores corresponding to the context combos if they exist?}

\item{run_file}{Name of file that contains the targets stores and code to create them corresponding to
the current project.}

\item{current_store_base, upstream_store_base}{Base directory for the current and upstream targets stores,
e.g. "../../out" or "projects/data".}

\item{settings_file}{File path of yaml file containing a list of default contexts (or 'scales') usually from
"settings/scales.yaml".
The contexts list must contain extent, grain & optionally aoi as first list elements, with secondary lists of
vector, filt_col, filt_level, buffer, extent_time, region_taxa under extent,
res_x, res_y, res_time, taxonomic under grain, and
vector, filt_col, filt_level, buffer under aoi.}

\item{list_name}{Name of the list in the settings file that contains the list of relevant contexts if it is
contained within another list. Use NULL if the extent, grain and aoi elements are primary and
not contained in another list.}

\item{run_all_combos}{List of specific settings vectors to vary the default settings with and
run all combinations of stores for. These are currently limited to extent_time, taxonomic grain & filt_level settings.}

\item{upstream_ext}{Character vector of upstream temporal extents corresponding to
the 'extent_time' setting in \code{settings$extent}, e.g. c("P10Y", "P20Y", "P30Y", "P50Y", "P100Y").
Often the same for current and upstream, and usually sourced from run_all_combos.
Only needed if different to current project extent_time specified in \code{run_all_combos}.
Numbers must be preceded by 'P' and followed by 'D', 'M', or 'Y'.}

\item{upstream_tax_grains}{Character vector of taxonomic grains for the current & upstream
projects, e.g. c("species", "subspecies"). Often the same for current and upstream, and
usually sourced from run_all_combos. Only needed if different to current project taxonomic grain/s
specified in \code{run_all_combos}.}

\item{current_lev, upstream_lev}{Vectors of current & upstream filter levels corresponding to the 'filt_level' setting
in \code{settings$grain}. Use NULL for upstream where the upstream project does not have an aoi setting.}

\item{lev_all_type}{Type of method used to determine 'all' values for \code{lev}.
Either, 'all_in_vec' which retrieves all levels in the column in the vector corresponding to those specified by
\code{settings$aoi$filt_col} and \code{settings$aoi$vector} respectively,
or 'already_run', which retrieves levels from existing store directories that have already been run,
i.e. those that are available.}

\item{current_aoi_setting, upstream_aoi_setting}{Names of the 'aoi' type settings for the current and
upstream projects containing the list of aoi related measures (vector, filt_col, filt_level, buffer),
e.g. sometimes referred to as 'region' instead of 'aoi'. Set to NULL if there is no aoi setting.}

\item{data_dir}{Path to data directory for sourcing the vector with the filter levels if \code{lev_all_type} = 'all_in_vec'.}

\item{current_region_taxa, upstream_region_taxa}{Logical. Is there a region taxa setting for the current and
upstream projects? Used to enable searching for track files in stores that have a region taxa setting and
those that don't.}

\item{current_proj}{Current project name in which running this function and running multiple outputs,
e.g. 'envRegCont'.}

\item{upstream_proj}{Upstream project name required for the current project,
i.e. the precursor project that contains the combos of settings context directories within them,
e.g. 'envRange'.}
}
\value{
Executes the run file for all the combinations of settings specified.

Executes the run file for all the combinations of settings specified.
}
\description{
Run all targets stores for multiple settings contexts

Run all targets stores for multiple settings contexts
}
