% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_dir.R
\name{store_dir}
\alias{store_dir}
\title{Generate a path to a storage directory}
\usage{
store_dir(
  set_list,
  base_dir = fs::path("..", "..", "out"),
  project = basename(here::here())
)
}
\arguments{
\item{set_list}{Named, nested list. Passed to \code{set_list} argument of
\code{envFunc::name_env_out()}}

\item{base_dir}{Character. Path to outputs store.}

\item{project}{Character. Name of the (usually Rstudio) project.}
}
\value{
Full path to a store directory for a project
}
\description{
Generate a path to a storage directory
}
