% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_store_clean.R
\name{summarise_store_clean}
\alias{summarise_store_clean}
\title{Summarise a cleaning workflow from a targets store}
\usage{
summarise_store_clean(
  store = ".",
  prefix = "bio_",
  excludes = c("path", "file", "env", "rich"),
  extras = c("clean_end"),
  site_cols = c("cell_lat", "cell_long"),
  visit_cols = c("cell_lat", "cell_long", "year"),
  taxa_cols = "taxa"
)
}
\arguments{
\item{store}{path to a \href{https://books.ropensci.org/targets/}{targets} store
(i.e. the \code{store} argument to \code{targets::tar_meta()})}

\item{prefix}{Object name prefixes. Objects within the \code{store} that match
this will be summarised}

\item{excludes}{Any object that matches any regex provided in \code{excludes} will
not be summarised}

\item{extras}{Any object that matches any regex provided in \code{extras} will be
summarised}

\item{site_cols, visit_cols, taxa_cols}{Columns within objects identifying
sites, visits and taxa respectively}
}
\description{
Summarise a cleaning workflow from a targets store
}
