% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_store_object.R
\name{summarise_store_object}
\alias{summarise_store_object}
\title{Summarise an object from a cleaning workflow from a targets store}
\usage{
summarise_store_object(
  store,
  object,
  prefix = "bio_",
  site_cols = c("lat", "long", "cell_lat", "cell_long"),
  visit_cols = c("lat", "long", "cell_lat", "cell_long", "year"),
  taxa_cols = c("original_name", "taxa"),
  rmd_dir = NULL,
  deps = NULL,
  use_arrow = TRUE,
  prep_filter_sf = TRUE,
  filter_sf_round = 3
)
}
\arguments{
\item{store}{Path to store containing \code{object}}

\item{object}{Name of \code{tars} element to summarise}

\item{prefix}{Character. e.g. \code{bio_} or \code{flor_}}

\item{site_cols, visit_cols, taxa_cols}{Columns within objects identifying
sites, visits and taxa respectively}

\item{rmd_dir}{What directory is the Rmd associated with each object found?
If left as \code{NULL}, will use \code{tars_name}.}

\item{deps}{Dependencies that, if updated, should invalidate the output.}

\item{use_arrow}{Logical. Use \code{arrow::open_dataset()} instead of
\code{tar_read_raw} on store objects. Saves memory, but requires targets to be
saved as parquets (with no file extension).}

\item{prep_filter_sf}{Logical. Prepare site-only (lat, long) data for
use in 'filter_sf' summary maps?}

\item{filter_sf_round}{Numeric. \code{digits} argument of \code{base::round()} for
use in rounding \code{site_cols}. Ignored if \code{prep_filter_sf} is not
\code{TRUE}.}
}
\value{
tibble
}
\description{
Returns a single row tibble that is equivalent to the multiple row tibble
returned by \code{envTargets::summarise_store_data()}. \code{summarise_store_object}
enables mapping over the objects in a targets workflow (potentially making
use of parallel processing to acheive the same result as
\code{summarise_store_data()}). Also see \code{envTargets::collect_clean_objects()} for
help creating a data frame to map over.
}
