% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tars.R
\name{write_tars}
\alias{write_tars}
\title{Write tars as yaml}
\usage{
write_tars(tars)
}
\arguments{
\item{tars}{Tars list}
}
\value{
\code{NULL}. Two \verb{_targets.yaml} files written
}
\description{
Writes a \code{tars} object to \code{here::here("_targets.yaml")} (i.e. the script
directory) and \code{fs::path(dirname(tars[[1]]$store), "_targets.yaml")} (i.e.
the store directory).
}
