% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_toi.R
\name{collect_toi}
\alias{collect_toi}
\title{Create a vector of taxa of interest (toi) from yaml files}
\usage{
collect_toi(
  dir = "settings",
  toi_regex = "\\\\/toi",
  environ_df = NULL,
  environ_remove = "marine",
  other_taxa = NULL
)
}
\arguments{
\item{dir}{Directory in which \code{toi_regex} files will be found}

\item{toi_regex}{Regular expression that will pick up any yaml files
containing the taxa of interest.}

\item{environ_df}{Dataframe of environments (e.g. terrestrial, marine).}

\item{environ_remove}{Character. Vector of environments to remove.}

\item{other_taxa}{Character. Vector of taxa used to limit the resulting
vector. i.e. only taxa that appear in \code{other_taxa} will make it into the
resulting vector.}
}
\value{
Character vector
}
\description{
Usually used to create a vector over which tarchetypes::tar_map is run
}
