% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tiles.R
\name{combine_tiles}
\alias{combine_tiles}
\title{Combine predicted tiles into a single raster}
\usage{
combine_tiles(predicted_tiles, out_file, ...)
}
\arguments{
\item{predicted_tiles}{Paths to predicted tiles, usually from
\code{predict_terra_tiles()}.}

\item{out_file}{Character. Name of the saved file.}

\item{...}{Passed to \code{terra::merge()}. e.g. \code{overwrite = TRUE}; or
\code{wopt = list(datatype = "INT1U")}.}
}
\value{
\code{out_file}
}
\description{
Combine predicted tiles into a single raster
}
