% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_tar_option_set.R
\name{env_tar_option_set}
\alias{env_tar_option_set}
\title{Set options for a targets workflow}
\usage{
env_tar_option_set(
  script,
  tars = yaml::read_yaml("_targets.yaml"),
  use_cores = envFunc::use_cores(absolute_max =
    yaml::read_yaml("settings/setup.yaml")$max_cores),
  pacs = yaml::read_yaml("settings/packages.yaml")[[script]]
)
}
\arguments{
\item{script}{Character. Which \code{tars} script is being run?}

\item{tars}{List, usually from \code{envTargets::make_tars()}}

\item{use_cores}{How many cores to use for \href{https://books.ropensci.org/targets/performance.html#worker-storage}{distributed computing}}

\item{pacs}{Character. Packages required for \code{script}}
}
\value{
Sets options for a targets workflow
}
\description{
Also enables logging and metric
\href{https://wlandau.github.io/crew/articles/logging.html#logging-worker-processes}{collection}.
See \code{envTargets::find_log_errors()} for using the logs.
}
