% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_tar_option_set.R
\name{env_tar_option_set}
\alias{env_tar_option_set}
\title{Set options for a targets workflow}
\usage{
env_tar_option_set(
  script,
  tars = yaml::read_yaml("_targets.yaml"),
  logs = TRUE,
  metrics = FALSE,
  controller_local_args = list(workers = envFunc::use_cores(absolute_max =
    yaml::read_yaml("settings/setup.yaml")$max_cores), name = "main_controller",
    crashes_max = 1L),
  tar_option_set_args = list(packages =
    yaml::read_yaml("settings/packages.yaml")[[script]], controller =
    get(controller_local_args$name))
)
}
\arguments{
\item{script}{Character. Which \code{tars} script is being run?}

\item{tars}{List, usually from \code{envTargets::make_tars()}}

\item{logs, metrics}{Logical. \href{https://wlandau.github.io/crew/articles/logging.html#logging-worker-processes}{Log worker processes}?}

\item{controller_local_args}{Named list of \code{crew::controller()} arguments}

\item{tar_options_set_args}{Named list of \code{crew::controller()} arguments}
}
\value{
Sets options for a targets workflow
}
\description{
Optionally enable logging and metric
\href{https://wlandau.github.io/crew/articles/logging.html#logging-worker-processes}{collection}.
See \code{envTargets::find_log_errors()} for using the logs.
}
