% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_env_stack.R
\name{make_env_stack}
\alias{make_env_stack}
\title{Make a spatRaster from env_df}
\usage{
make_env_stack(env_df)
}
\arguments{
\item{env_df}{tibble. Probably a result of \code{prepare_env()}.}
}
\value{
spatRaster
}
\description{
Make a spatRaster from env_df
}
