% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tile_extents.R
\name{make_tile_extents}
\alias{make_tile_extents}
\title{Make tile (extents) for raster split, apply, combine}
\usage{
make_tile_extents(
  base_grid_path,
  aoi,
  tile_size = 1e+05,
  min_tiles = 4,
  add_hectares = TRUE
)
}
\arguments{
\item{base_grid_path}{Character. Path to a template raster.}

\item{aoi}{Optional sf object. Used to define the outer extent of tiles}

\item{tile_size}{Numeric. Tile size in units of the raster at \code{base_grid_path}}

\item{min_tiles}{Numeric. Minimum number of tiles to return. If \code{tile_size}
returns less than \code{min_tiles}, then \code{min_tiles} will be used instead of
\code{tile_size}}

\item{add_hectares}{Logical. Append the hectares of each tile? Ignored if
the raster at \code{base_grid_path} is not projected.}
}
\value{
tibble of extents
}
\description{
Make tile (extents) for raster split, apply, combine
}
