% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_terra_tiles.R
\name{predict_terra_tiles}
\alias{predict_terra_tiles}
\title{Predict a model across tiles}
\usage{
predict_terra_tiles(
  extent_df,
  predict_stack,
  model,
  terra_options = NULL,
  out_dir,
  force_new = TRUE,
  load_packages = NULL,
  ...
)
}
\arguments{
\item{extent_df}{Tibble from \code{make_tile_extents()}.}

\item{predict_stack}{SpatRaster, possibly from \code{make_env_stack()}.}

\item{model}{Model object with a predict method.}

\item{terra_options}{List of options to be passed to \code{terra::terraOptions()}.
e.g. \code{list(memfrac = 0.5)}.}

\item{out_dir}{Character. Name of directory into which predicted tiles will
be saved.}

\item{force_new}{Logical. If TRUE, predict tile even if it already exists.
Leave as \code{TRUE} to let targets decide if the tile needs to be rewritten.}

\item{load_packages}{Character. Any packages to load before predict.}

\item{...}{Passed to \code{terra::predict()}. Object, model and filename arguments
are already passed so are not needed here. e.g. \code{na.rm = TRUE};
\code{overwrite = TRUE}; or \code{wopt = list(datatype = "INT1U")}}
}
\value{
Path to the tile predicted tile. Writes tile into \code{out_dir}.
}
\description{
Predict a model across tiles
}
