% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_env.R
\name{prepare_env}
\alias{prepare_env}
\title{Prepare a tibble of environmental layers from settings}
\usage{
prepare_env(
  set_list,
  reg_exp = "\\\\.tif",
  base_dir = if (Sys.info()["sysname"] == "Windows") "I:" else fs::path("/mnt/envcube",
    ""),
  create_short_desc = TRUE,
  is_static = FALSE,
  max_date_only = TRUE,
  ...
)
}
\arguments{
\item{set_list}{list. usually from yaml::read_yaml("settings/setup.yaml")}

\item{reg_exp}{Character. Used to limit returned files.}

\item{base_dir}{Character. Passed to \code{base_dir} argument of
\code{envFunc::name_env_out()}}

\item{create_short_desc}{Logical. Create a short description to use as a
unique 'name'. Requires columns 'name' and 'func'.}

\item{is_static}{Logical. If the \code{start_date} field is 'static' set this to
TRUE.}

\item{max_date_only}{Logical. Only used if \code{is_static} is TRUE. Filter
rows duplicated by \code{layer} and \code{func} to the latest \code{start_date}.}

\item{...}{Passed to envRaster::name_env_tif}
}
\value{
tibble of raster paths and meta data parsed into columns
}
\description{
'Seasons' functionality removed 2025-09-05. If seasons are needed, left join
the result of \code{prepare_env()} to the result of \verb{make_seasons()``$seasons}.
}
