% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_env.R
\name{prepare_env}
\alias{prepare_env}
\title{Prepare a tibble of environmental layers from settings}
\usage{
prepare_env(
  set_list,
  reg_exp = "\\\\.tif",
  base_year = as.numeric(format(Sys.Date(), "\%Y")) - 10,
  ...
)
}
\arguments{
\item{set_list}{list. usually from yaml::read_yaml("settings/setup.yaml")}

\item{reg_exp}{Character. Used to limit returned files.}

\item{base_year}{Numeric. Default is
\code{as.numeric(format(Sys.Date(), "\%Y")) - 10}}

\item{...}{Passed to envRaster::name_env_tif}
}
\value{
tibble
}
\description{
Prepare a tibble of environmental layers from settings
}
