% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_with_deps.R
\name{render_with_deps}
\alias{render_with_deps}
\title{Render bookdown and force Rmd file dependencies}
\usage{
render_with_deps(
  input_directory = ".",
  deps,
  output_directory = NULL,
  clean_up = TRUE
)
}
\arguments{
\item{input_directory}{Path in which \code{index.Rmd} can be found}

\item{output_directory}{Directory path. Where should the final results be
saved (this is usually not a targets store).}

\item{clean_up}{Logical. Delete intermediary \emph{and} \code{output_directory} \emph{before}
knit and delete intermediary files \emph{after} knit.}
}
\value{
The target is just the path \code{output_directory}. Files needed for the
book are written into \code{output_directory}.
}
\description{
Enables a bookdown 'target' within a \href{https://books.ropensci.org/targets/}{targets workflow}.
I was unable to get \href{https://mps9506.r-universe.dev/bookdowntargets/doc/manual.html}{bookdowntargets}
to work without error. That may be a better way to achieve this. Currently
(20250520) \code{deps} still need to be provided inside bookdown scripts (i.e. providing
\code{deps} to the function simply tells targets when the book needs to be remade,
it does not make those \code{deps} available inside the knit.)
}
\details{
Based on code found here: https://github.com/shirdekel/phd_thesis/commit/cc66e1b9d9305a4b21e8489836545ecc4475f9ee
}
\author{
Shir Dekel (modified by nw)
}
