% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_with_deps.R
\name{render_with_deps}
\alias{render_with_deps}
\title{Render bookdown and force Rmd file dependencies}
\usage{
render_with_deps(
  input_directory = ".",
  deps,
  remove_main = TRUE,
  clean_out_dir = TRUE,
  clean_up = TRUE
)
}
\arguments{
\item{input_directory}{Path in which \code{index.Rmd} can be found.}

\item{deps}{Any dependencies for the report. Note that these are not made
available within the knit, they just trigger the target to run.}

\item{remove_main}{Logical. Delete _main.Rmd froom the input directory before
knit?}

\item{clean_out_dir}{Logical. Delete the book before knit? Useful if chapter
names are changed, leaving orphaned \code{.html} files in the output directory.}

\item{clean_up}{Logical. Delete intermediary files after knit?}
}
\value{
Return value is just the \code{output_dir} specified in \verb{_bookdown.yaml},
or, if that is not specified, the value of \code{input_directory}. Files needed
for the book are written into the output directory.
}
\description{
Moving the report to the targets store relies on output_dir being specified
in the _bookdown.yaml file, usually written within report_prep.R
}
\details{
Enables a bookdown 'target' within a \href{https://books.ropensci.org/targets/}{targets workflow}.
I was unable to get \href{https://mps9506.r-universe.dev/bookdowntargets/doc/manual.html}{bookdowntargets}
to work without error. That may be a better way to achieve this. Currently
(20250520) \code{deps} still need to be provided inside bookdown scripts (i.e. providing
\code{deps} to the function simply tells targets when the book needs to be remade,
it does not make those \code{deps} available inside the knit.)

Based on code found here: https://github.com/shirdekel/phd_thesis/commit/cc66e1b9d9305a4b21e8489836545ecc4475f9ee
}
\author{
Shir Dekel (modified by nw)
}
