% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{Run all targets stores for multiple settings contexts}
\usage{
run_all(
  scales_yaml = "scales.yaml",
  combos_df,
  current_proj = basename(here::here()),
  upstream_proj,
  upstream_track_file = "_targets.yaml",
  current_track_file = "_targets.yaml",
  force_new = TRUE,
  run_file = "run.R",
  current_store_base = fs::path("..", "..", "out"),
  upstream_store_base = fs::path("..", "..", "out")
)
}
\arguments{
\item{scales_yaml}{Character. Name of the \code{scales.yaml} file to obtain settings contexts.
Must be the same in both the current and upstream projects.}

\item{combos_df}{Data frame of combinations of the settings to be varied, where the columns are
the setting names and the rows are each combination of different values for each setting,
e.g. as produced by envTargets::make_context_combos. The setting names must correspond to those in
the settings for the \code{current_proj} element in the \code{scales_yaml}.}

\item{current_proj}{Current project name in which running this function and running multiple outputs,
e.g. 'envRegCont'.}

\item{upstream_proj}{Upstream project name required for the current project,
i.e. the precursor project that contains the combos of settings context directories within them,
e.g. 'envRange'.}

\item{current_track_file, upstream_track_file}{Names of files to use for tracking if a store
relating to a context combo has been run. Usually one of the last files created in the project/store,
and or one used downstream, e.g. 'grd_files' in the envRange project as the upstream file, and
'reg_cont_tbl_tidy' in envRegCont if that project was the current. Otherwise, defaults to the
'_targets.yaml', as an indication the store has been run (although possibly not fully).}

\item{force_new}{Logical. Force new runs of stores corresponding to the context combos if they exist?}

\item{run_file}{Name of file that contains the targets stores and code to create them corresponding to
the current project.}

\item{current_store_base, upstream_store_base}{Base directory for the current and upstream targets stores,
e.g. "../../out" or "projects/data".}
}
\value{
Executes the run file for all the combinations of settings specified.
}
\description{
Run all targets stores for multiple settings contexts
}
