% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_progress.R
\name{sdm_progress}
\alias{sdm_progress}
\title{Check completion of an envSDMs run}
\usage{
sdm_progress(
  sdm_store = tars$sdm$store,
  grain = "fine",
  find_progress = c("boundary", "prep", "tune", "full_run", "pred", "thresh"),
  taxa = NULL
)
}
\arguments{
\item{grain}{"fine" or "coarse" - which output directory to search}

\item{find_progress}{Which files to check for (skipping boundary can )}

\item{taxa}{Optionally check for only specified taxa. If NULL (default), will \code{tar_read} toi from the setup store.}
}
\value{
Tibble with columns 'toi', each of \code{find_progress}, and 'finished' (which checks for a pred .log file)
}
\description{
Checks for the existence of files in relevant taxa's output directories (or reads the 'boundary' target).
}
