% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_store_data.R
\name{summarise_store_data}
\alias{summarise_store_data}
\title{Summarise a cleaning workflow from a targets store}
\usage{
summarise_store_data(
  tars = NULL,
  tars_name = "clean",
  prefix = "bio_",
  excludes = c("path", "dir", "file", "env", "rich"),
  extras = c("clean_end"),
  site_cols = c("cell_lat", "cell_long"),
  visit_cols = c("cell_lat", "cell_long", "year"),
  taxa_cols = "taxa",
  rmd_dir = NULL,
  deps = NULL,
  use_arrow = TRUE
)
}
\arguments{
\item{tars}{Named list of target stores and scripts}

\item{tars_name}{Name of \code{tars} element to summarise}

\item{prefix}{Object name prefixes. Objects within \code{tars[[tars_name]]$store}
that match this will be summarised}

\item{excludes}{Any object that matches any regex provided in \code{excludes} will
not be summarised}

\item{extras}{Any object that matches any regex provided in \code{extras} will be
summarised}

\item{site_cols, visit_cols, taxa_cols}{Columns within objects identifying
sites, visits and taxa respectively}

\item{rmd_dir}{What directory is the Rmd associated with each object found?
If left as \code{NULL}, will use \code{tars_name}.}

\item{deps}{Dependencies that, if updated, should invalidate the output.}

\item{use_arrow}{Use \code{arrow::open_dataset()} instead of \code{tar_read_raw} on store objects. Saves memory, but requires targets to be saved as parquets (with no file extension).}
}
\description{
Summarise a cleaning workflow from a targets store
}
